+++ jcm/physics/forcing/speedy_forcing.py
